migrate StrEnum to main area
