+++ torchvision/prototype/features/_bounding_box.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import Any, Tuple, Union, Optional
 
 import torch
-from torchvision.prototype.utils._internal import StrEnum
+from torchvision._utils import StrEnum
 
 from ._feature import _Feature
