+++ torchvision/prototype/utils/_internal.py
@@ -1,6 +1,5 @@
 import collections.abc
 import difflib
-import enum
 import functools
 import inspect
 import io
@@ -31,7 +30,6 @@ import numpy as np
 import torch
 
 __all__ = [
-    "StrEnum",
     "sequence_to_str",
     "add_suggestion",
     "FrozenMapping",
@@ -43,26 +41,6 @@ __all__ = [
     "apply_recursively",
     "query_recursively",
 ]
-
-
-class StrEnumMeta(enum.EnumMeta):
-    auto = enum.auto
-
-    def from_str(self, member: str):
-        try:
-            return self[member]
-        except KeyError:
-            raise ValueError(
-                add_suggestion(
-                    f"Unknown value '{member}' for {self.__name__}.",
-                    word=member,
-                    possibilities=list(self.__members__.keys()),
-                )
-            ) from None
-
-
-class StrEnum(enum.Enum, metaclass=StrEnumMeta):
-    pass
 
 
 def sequence_to_str(seq: Sequence, separate_last: str = "") -> str:
